refactor: rename auth modules
