+++ packages/routed/lib/src/auth/providers.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:http/http.dart' as http;
-import 'package:routed/src/auth/auth_models.dart';
+import 'package:routed/src/auth/models.dart';
 import 'package:routed/src/auth/oauth.dart';
 import 'package:routed/src/context/context.dart';
